Building_End_to_End_LLM_and_Large_image_model_application_using_Gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
ibm-watson-machine-learning
langchain-community